% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{being_processed}
\alias{being_processed}
\title{being_processed}
\usage{
being_processed(dat)
}
\arguments{
\item{dat}{Output from \code{determine_new_datasets}}
}
\value{
Updated dat
}
\description{
List of EBI datasets that are currently being processed
}
