% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{create_build_reference}
\alias{create_build_reference}
\title{Create dataset of some hm3 SNPs and their build positions}
\usage{
create_build_reference()
}
\value{
saves build_ref object
}
\description{
Create dataset of some hm3 SNPs and their build positions
}
