% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.r
\name{Dataset}
\alias{Dataset}
\title{Object that downloads, develops and uploads GWAS summary datasets for IEU OpenGWAS database}
\description{
Object that downloads, develops and uploads GWAS summary datasets for IEU OpenGWAS database

Object that downloads, develops and uploads GWAS summary datasets for IEU OpenGWAS database
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-check_id}{\code{Dataset$check_id()}}
\item \href{#method-finalize}{\code{Dataset$finalize()}}
\item \href{#method-delete_wd}{\code{Dataset$delete_wd()}}
\item \href{#method-set_wd}{\code{Dataset$set_wd()}}
\item \href{#method-determine_columns}{\code{Dataset$determine_columns()}}
\item \href{#method-format_dataset}{\code{Dataset$format_dataset()}}
\item \href{#method-get_required_fields}{\code{Dataset$get_required_fields()}}
\item \href{#method-collect_metadata}{\code{Dataset$collect_metadata()}}
\item \href{#method-write_metadata}{\code{Dataset$write_metadata()}}
\item \href{#method-api_metadata_upload}{\code{Dataset$api_metadata_upload()}}
\item \href{#method-api_metadata_check}{\code{Dataset$api_metadata_check()}}
\item \href{#method-api_metadata_delete}{\code{Dataset$api_metadata_delete()}}
\item \href{#method-api_gwasdata_upload}{\code{Dataset$api_gwasdata_upload()}}
\item \href{#method-api_gwasdata_check}{\code{Dataset$api_gwasdata_check()}}
\item \href{#method-api_gwasdata_delete}{\code{Dataset$api_gwasdata_delete()}}
\item \href{#method-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(filename, wd = tempdir(), igd_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{igd_id}}{Defaults to "ebi-a-<ebi_id>"}

\item{\code{ebi_id}}{e.g. GCST004426}

\item{\code{wd=tempdir()}}{<what param does>}

\item{\code{ftp_path=NULL}}{<what param does>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new ObtainEbiDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_id"></a>}}
\if{latex}{\out{\hypertarget{method-check_id}{}}}
\subsection{Method \code{check_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$check_id(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_wd"></a>}}
\if{latex}{\out{\hypertarget{method-delete_wd}{}}}
\subsection{Method \code{delete_wd()}}{
delete working directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$delete_wd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_wd"></a>}}
\if{latex}{\out{\hypertarget{method-set_wd}{}}}
\subsection{Method \code{set_wd()}}{
set working directory (creates)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_wd(wd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wd}}{working directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-determine_columns"></a>}}
\if{latex}{\out{\hypertarget{method-determine_columns}{}}}
\subsection{Method \code{determine_columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$determine_columns(params, nrows = 100, gwas_file = self$filename, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-format_dataset}{}}}
\subsection{Method \code{format_dataset()}}{
format dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$format_dataset(
  gwas_file = self$filename,
  gwas_out = file.path(self$wd, "format.txt.gz"),
  params = self$params,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dl=self$dl}}{<what param does>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_required_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-collect_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-collect_metadata}{}}}
\subsection{Method \code{collect_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$collect_metadata(metadata, igd_id = self$igd_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-write_metadata}{}}}
\subsection{Method \code{write_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$write_metadata(
  metadata = self$metadata,
  datainfo = self$datainfo,
  outdir = self$wd
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_metadata_upload"></a>}}
\if{latex}{\out{\hypertarget{method-api_metadata_upload}{}}}
\subsection{Method \code{api_metadata_upload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$api_metadata_upload(
  metadata = self$metadata,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_metadata_check"></a>}}
\if{latex}{\out{\hypertarget{method-api_metadata_check}{}}}
\subsection{Method \code{api_metadata_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$api_metadata_check(
  id = self$igd_id,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_metadata_delete"></a>}}
\if{latex}{\out{\hypertarget{method-api_metadata_delete}{}}}
\subsection{Method \code{api_metadata_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$api_metadata_delete(
  id = self$igd_id,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_gwasdata_upload"></a>}}
\if{latex}{\out{\hypertarget{method-api_gwasdata_upload}{}}}
\subsection{Method \code{api_gwasdata_upload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$api_gwasdata_upload(
  datainfo = self$datainfo,
  gwasfile = self$gwas_out,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_gwasdata_check"></a>}}
\if{latex}{\out{\hypertarget{method-api_gwasdata_check}{}}}
\subsection{Method \code{api_gwasdata_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$api_gwasdata_check(
  id = self$igd_id,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api_gwasdata_delete"></a>}}
\if{latex}{\out{\hypertarget{method-api_gwasdata_delete}{}}}
\subsection{Method \code{api_gwasdata_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$api_gwasdata_delete(
  id = self$igd_id,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
