% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{determine_build}
\alias{determine_build}
\title{Determine build based on reference dataset}
\usage{
determine_build(rsid, chr, pos, build = c(37, 38, 36))
}
\arguments{
\item{rsid}{rsid}

\item{chr}{chr}

\item{pos}{pos}

\item{build}{Builds to try e.g. c(37,38,36)}
}
\value{
build if detected, or dataframe of matches if not
}
\description{
Determine build based on reference dataset
}
