% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{determine_build_biomart}
\alias{determine_build_biomart}
\title{Determines which build a set of SNps are on}
\usage{
determine_build_biomart(rsid, chr, pos, build = c(37, 38, 36))
}
\arguments{
\item{rsid}{rsid}

\item{chr}{chr}

\item{pos}{pos}

\item{build}{Builds to try e.g. c(37,38,36)}
}
\value{
build if detected, or dataframe of matches if not
}
\description{
Determines which build a set of SNps are on
}
