% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{determine_build_position}
\alias{determine_build_position}
\title{Determine build just from positions}
\usage{
determine_build_position(pos, build = c(37, 38, 36))
}
\arguments{
\item{pos}{Vector of positions}

\item{build}{c(37,38,36)}
}
\value{
build or if not determined then dataframe
}
\description{
A bit sketchy but computationally fast - just assumes that there will be at least 50x more position matches in the true build than either of the others.
}
