% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{determine_new_datasets}
\alias{determine_new_datasets}
\title{determine_new_datasets}
\usage{
determine_new_datasets(
  ebi_ftp_url = options()$ebi_ftp_url,
  blacklist = NULL,
  exclude_multi_datasets = TRUE
)
}
\arguments{
\item{ebi_ftp_url}{FTP url default=options()$ebi_ftp_url}

\item{blacklist}{List of EBI datasets to ignore default=NULL}

\item{exclude_multi_datasets}{If a EBI ID has more than one dataset then should it be ignored}
}
\value{
data frame
}
\description{
Figure out which datasets are not present in the database
}
