% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{ebi_datasets}
\alias{ebi_datasets}
\title{ebi_datasets}
\usage{
ebi_datasets(ebi_ftp_url = options()$ebi_ftp_url)
}
\arguments{
\item{ebi_ftp_url}{EBI FTP default=options()$ebi_ftp_url}
}
\value{
data frame
}
\description{
Convert output from listftp into something that is easier to read
}
