% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{get_ftp_path}
\alias{get_ftp_path}
\title{Get harmonised file for specific EBI ID}
\usage{
get_ftp_path(ebi_id, ebi_ftp_url = options()$ebi_ftp_url)
}
\arguments{
\item{ebi_id}{EBI ID e.g. GCST000879}

\item{ebi_ftp_url=options()$ebi_ftp_url}{<what param does>}
}
\value{
ftp path (excluding server)
}
\description{
Get harmonised file for specific EBI ID
}
