% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{liftover_gwas}
\alias{liftover_gwas}
\title{Liftover GWAS positions}
\usage{
liftover_gwas(dat, build = c(37, 38, 36), to = 37)
}
\arguments{
\item{build}{Builds to try e.g. c(37,38,36)}

\item{rsid}{rsid}

\item{chr}{chr}

\item{pos}{pos}
}
\value{
NULL if already build 37 or dataframe of b37 positions
}
\description{
Determine GWAS build and liftover to b37
}
