% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{listftp}
\alias{listftp}
\title{listftp}
\usage{
listftp(url = options()$ebi_ftp_url, recursive = TRUE)
}
\arguments{
\item{url}{FTP url to look up}

\item{recursive}{If false then just the top directory, otherwise list recursively}
}
\value{
Vector of paths
}
\description{
List all files on the EBI FTP server
}
