% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain.r
\name{ObtainEBIDataset}
\alias{ObtainEBIDataset}
\title{Obtain EBI Dataset class}
\description{
Obtain EBI Dataset class

Obtain EBI Dataset class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObtainEBIDataset$new()}}
\item \href{#method-delete_wd}{\code{ObtainEBIDataset$delete_wd()}}
\item \href{#method-set_wd}{\code{ObtainEBIDataset$set_wd()}}
\item \href{#method-download_dataset}{\code{ObtainEBIDataset$download_dataset()}}
\item \href{#method-format_dataset}{\code{ObtainEBIDataset$format_dataset()}}
\item \href{#method-organise_metadata}{\code{ObtainEBIDataset$organise_metadata()}}
\item \href{#method-write_metadata}{\code{ObtainEBIDataset$write_metadata()}}
\item \href{#method-upload_metadata}{\code{ObtainEBIDataset$upload_metadata()}}
\item \href{#method-upload_check}{\code{ObtainEBIDataset$upload_check()}}
\item \href{#method-upload_delete}{\code{ObtainEBIDataset$upload_delete()}}
\item \href{#method-upload_gwas}{\code{ObtainEBIDataset$upload_gwas()}}
\item \href{#method-clone}{\code{ObtainEBIDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$new(ebi_id, wd = tempdir(), ftp_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id}}{e.g. GCST004426}

\item{\code{wd=tempdir()}}{<what param does>}

\item{\code{ftp_path=NULL}}{<what param does>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new ObtainEBIDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_wd"></a>}}
\if{latex}{\out{\hypertarget{method-delete_wd}{}}}
\subsection{Method \code{delete_wd()}}{
delete working directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$delete_wd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_wd"></a>}}
\if{latex}{\out{\hypertarget{method-set_wd}{}}}
\subsection{Method \code{set_wd()}}{
set working directory (creates)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$set_wd(wd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wd}}{working directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-download_dataset}{}}}
\subsection{Method \code{download_dataset()}}{
Download
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$download_dataset(
  ftp_path = self$ftp_path,
  ftp_url = options()$ebi_ftp_url,
  outdir = self$wd
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ftp_path=self$ftp_path}}{<what param does>}

\item{\code{ftp_url=options()$ebi_ftp_url}}{<what param does>}

\item{\code{outdir=self$wd}}{<what param does>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-format_dataset}{}}}
\subsection{Method \code{format_dataset()}}{
format dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$format_dataset(dl = self$dl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dl=self$dl}}{<what param does>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-organise_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-organise_metadata}{}}}
\subsection{Method \code{organise_metadata()}}{
Download and parse metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$organise_metadata(
  ebi_id = self$ebi_id,
  or_flag = self$or_flag,
  igd_id = NULL,
  units = NULL,
  sex = "NA",
  category = "NA",
  subcategory = "NA",
  build = "HG19/GRCh37",
  group_name = "public"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id=self$ebi_id}}{<what param does>}

\item{\code{or_flag=self$or_flag}}{<what param does>}

\item{\code{igd_id=NULL}}{<what param does>}

\item{\code{units=NULL}}{<what param does>}

\item{\code{sex="NA"}}{<what param does>}

\item{\code{category="NA"}}{<what param does>}

\item{\code{subcategory="NA"}}{<what param does>}

\item{\code{build="HG19/GRCh37"}}{<what param does>}

\item{\code{group_name="public"}}{<what param does>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-write_metadata}{}}}
\subsection{Method \code{write_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$write_metadata(
  metadata = self$metadata,
  datainfo = self$datainfo,
  outdir = self$wd
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-upload_metadata}{}}}
\subsection{Method \code{upload_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$upload_metadata(
  metadata_file = self$metadata_file,
  datainfo_file = self$datainfo_file,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_check"></a>}}
\if{latex}{\out{\hypertarget{method-upload_check}{}}}
\subsection{Method \code{upload_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$upload_check(
  id = self$metadata$id,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_delete"></a>}}
\if{latex}{\out{\hypertarget{method-upload_delete}{}}}
\subsection{Method \code{upload_delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$upload_delete(
  id = self$metadata$id,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_gwas"></a>}}
\if{latex}{\out{\hypertarget{method-upload_gwas}{}}}
\subsection{Method \code{upload_gwas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$upload_gwas(
  datainfo_file = self$datainfo_file,
  gwasfile = self$datainfo$filename,
  access_token = ieugwasr::check_access_token()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEBIDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
