% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{determine_new_datasets}
\alias{determine_new_datasets}
\title{<brief desc>}
\usage{
determine_new_datasets(
  ebi_ftp_url = options()$ebi_ftp_url,
  blacklist = NULL,
  exclude_multi_datasets = TRUE
)
}
\arguments{
\item{ebi_ftp_url=options()$ebi_ftp_url}{<what param does>}

\item{blacklist=NULL}{<what param does>}
}
\value{
data frame
}
\description{
<full description>
}
