% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{ebi_datasets}
\alias{ebi_datasets}
\title{ebi_datasets}
\usage{
ebi_datasets(ebi_ftp_url = options()$ebi_ftp_url)
}
\arguments{
\item{ebi_ftp_url=options()$ebi_ftp_url}{<what param does>}
}
\value{

}
\description{
<full description>
}
