% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EbiDataset.r
\name{EbiDataset}
\alias{EbiDataset}
\title{Object that downloads, develops and uploads EBI dataset}
\description{
Object that downloads, develops and uploads EBI dataset

Object that downloads, develops and uploads EBI dataset
}
\section{Super class}{
\code{\link[GwasDataImport:Dataset]{GwasDataImport::Dataset}} -> \code{EbiDataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ebi_id}}{EBI ID to look for}

\item{\code{traitname}}{Name of trait}

\item{\code{ftp_path}}{Path to files in EBI FTP}

\item{\code{or_flag}}{TRUE/FALSE if had to convert OR to beta}

\item{\code{gwas_out1}}{Path to first look at EBI dataset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EbiDataset-new}{\code{EbiDataset$new()}}
\item \href{#method-EbiDataset-download_dataset}{\code{EbiDataset$download_dataset()}}
\item \href{#method-EbiDataset-format_ebi_dataset}{\code{EbiDataset$format_ebi_dataset()}}
\item \href{#method-EbiDataset-organise_metadata}{\code{EbiDataset$organise_metadata()}}
\item \href{#method-EbiDataset-pipeline}{\code{EbiDataset$pipeline()}}
\item \href{#method-EbiDataset-clone}{\code{EbiDataset$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwas_release"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_gwas_release'><code>GwasDataImport::Dataset$api_gwas_release()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwasdata_check"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_gwasdata_check'><code>GwasDataImport::Dataset$api_gwasdata_check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwasdata_delete"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_gwasdata_delete'><code>GwasDataImport::Dataset$api_gwasdata_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwasdata_upload"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_gwasdata_upload'><code>GwasDataImport::Dataset$api_gwasdata_upload()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_check"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_metadata_check'><code>GwasDataImport::Dataset$api_metadata_check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_delete"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_metadata_delete'><code>GwasDataImport::Dataset$api_metadata_delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_edit"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_metadata_edit'><code>GwasDataImport::Dataset$api_metadata_edit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_upload"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_metadata_upload'><code>GwasDataImport::Dataset$api_metadata_upload()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_qc_status"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_qc_status'><code>GwasDataImport::Dataset$api_qc_status()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_report"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-api_report'><code>GwasDataImport::Dataset$api_report()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="check_metadata"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-check_metadata'><code>GwasDataImport::Dataset$check_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="collect_metadata"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-collect_metadata'><code>GwasDataImport::Dataset$collect_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="delete_wd"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-delete_wd'><code>GwasDataImport::Dataset$delete_wd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="determine_columns"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-determine_columns'><code>GwasDataImport::Dataset$determine_columns()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="format_dataset"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-format_dataset'><code>GwasDataImport::Dataset$format_dataset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="get_gwasdata_fields"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-get_gwasdata_fields'><code>GwasDataImport::Dataset$get_gwasdata_fields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="get_metadata_fields"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-get_metadata_fields'><code>GwasDataImport::Dataset$get_metadata_fields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="is_new_id"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-is_new_id'><code>GwasDataImport::Dataset$is_new_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="se_from_bp"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-se_from_bp'><code>GwasDataImport::Dataset$se_from_bp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="set_wd"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-set_wd'><code>GwasDataImport::Dataset$set_wd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="view_metadata_options"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-view_metadata_options'><code>GwasDataImport::Dataset$view_metadata_options()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="write_metadata"><a href='../../GwasDataImport/html/Dataset.html#method-Dataset-write_metadata'><code>GwasDataImport::Dataset$write_metadata()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EbiDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-EbiDataset-new}{}}}
\subsection{Method \code{new()}}{
Initialise object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$new(
  ebi_id,
  wd = tempdir(),
  ftp_path = NULL,
  igd_id = paste0("ebi-a-", ebi_id),
  traitname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id}}{e.g. GCST005522}

\item{\code{wd}}{Directory in which to download and develop dataset. Default=tempdir(). Deleted automatically upon object removal}

\item{\code{ftp_path}}{Pre-specified path to data. Default=NULL}

\item{\code{igd_id}}{Defaults to "ebi-a-<ebi_id>"}

\item{\code{traitname}}{Option to provide traitname of dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new EbiDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EbiDataset-download_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-EbiDataset-download_dataset}{}}}
\subsection{Method \code{download_dataset()}}{
Download
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$download_dataset(
  ftp_path = self$ftp_path,
  outdir = self$wd,
  dl = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ftp_path}}{Pre-specified path to data. Default=self$ftp_path}

\item{\code{outdir}}{Default=self$wd}

\item{\code{ftp_url}}{Default=options()$ebi_ftp_url}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EbiDataset-format_ebi_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-EbiDataset-format_ebi_dataset}{}}}
\subsection{Method \code{format_ebi_dataset()}}{
organise data before formatting. This is slow but doesn't really matter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$format_ebi_dataset(
  filename = self$filename,
  output = file.path(self$wd, "step1.txt.gz")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Filename of GWAS dataset}

\item{\code{output}}{Where to save formatted dataset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EbiDataset-organise_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-EbiDataset-organise_metadata}{}}}
\subsection{Method \code{organise_metadata()}}{
Download and parse metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$organise_metadata(
  ebi_id = self$ebi_id,
  or_flag = self$or_flag,
  igd_id = self$igd_id,
  units = NULL,
  sex = "NA",
  category = "NA",
  subcategory = "NA",
  build = "HG19/GRCh37",
  group_name = "public",
  traitname = self$traitname
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id}}{Default=self$ebi_id}

\item{\code{or_flag}}{Default=self$or_flag}

\item{\code{igd_id}}{Default=NULL}

\item{\code{units}}{Default=NULL}

\item{\code{sex}}{Default="NA"}

\item{\code{category}}{Default="NA"}

\item{\code{subcategory}}{Default="NA"}

\item{\code{build}}{Default="HG19/GRCh37"}

\item{\code{group_name}}{Default="public"}

\item{\code{traitname}}{Default=self$traitname}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EbiDataset-pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-EbiDataset-pipeline}{}}}
\subsection{Method \code{pipeline()}}{
Once initialised this function will string together everything i.e. downloading, processing and uploading
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$pipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EbiDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EbiDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
