% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EbiDataset.r
\name{EbiDataset}
\alias{EbiDataset}
\title{Object that downloads, develops and uploads EBI dataset}
\description{
Object that downloads, develops and uploads EBI dataset

Object that downloads, develops and uploads EBI dataset
}
\section{Super class}{
\code{\link[GwasDataImport:Dataset]{GwasDataImport::Dataset}} -> \code{EbiDataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ebi_id}}{EBI ID to look for}

\item{\code{traitname}}{Name of trait}

\item{\code{ftp_path}}{Path to files in EBI FTP}

\item{\code{or_flag}}{TRUE/FALSE if had to convert OR to beta}

\item{\code{gwas_out1}}{Path to first look at EBI dataset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EbiDataset$new()}}
\item \href{#method-download_dataset}{\code{EbiDataset$download_dataset()}}
\item \href{#method-format_ebi_dataset}{\code{EbiDataset$format_ebi_dataset()}}
\item \href{#method-organise_metadata}{\code{EbiDataset$organise_metadata()}}
\item \href{#method-pipeline}{\code{EbiDataset$pipeline()}}
\item \href{#method-clone}{\code{EbiDataset$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwas_release">}\href{../../GwasDataImport/html/Dataset.html#method-api_gwas_release}{\code{GwasDataImport::Dataset$api_gwas_release()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwasdata_check">}\href{../../GwasDataImport/html/Dataset.html#method-api_gwasdata_check}{\code{GwasDataImport::Dataset$api_gwasdata_check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwasdata_delete">}\href{../../GwasDataImport/html/Dataset.html#method-api_gwasdata_delete}{\code{GwasDataImport::Dataset$api_gwasdata_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_gwasdata_upload">}\href{../../GwasDataImport/html/Dataset.html#method-api_gwasdata_upload}{\code{GwasDataImport::Dataset$api_gwasdata_upload()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_check">}\href{../../GwasDataImport/html/Dataset.html#method-api_metadata_check}{\code{GwasDataImport::Dataset$api_metadata_check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_delete">}\href{../../GwasDataImport/html/Dataset.html#method-api_metadata_delete}{\code{GwasDataImport::Dataset$api_metadata_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_edit">}\href{../../GwasDataImport/html/Dataset.html#method-api_metadata_edit}{\code{GwasDataImport::Dataset$api_metadata_edit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_metadata_upload">}\href{../../GwasDataImport/html/Dataset.html#method-api_metadata_upload}{\code{GwasDataImport::Dataset$api_metadata_upload()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_qc_status">}\href{../../GwasDataImport/html/Dataset.html#method-api_qc_status}{\code{GwasDataImport::Dataset$api_qc_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="api_report">}\href{../../GwasDataImport/html/Dataset.html#method-api_report}{\code{GwasDataImport::Dataset$api_report()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="collect_metadata">}\href{../../GwasDataImport/html/Dataset.html#method-collect_metadata}{\code{GwasDataImport::Dataset$collect_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="delete_wd">}\href{../../GwasDataImport/html/Dataset.html#method-delete_wd}{\code{GwasDataImport::Dataset$delete_wd()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="determine_columns">}\href{../../GwasDataImport/html/Dataset.html#method-determine_columns}{\code{GwasDataImport::Dataset$determine_columns()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="format_dataset">}\href{../../GwasDataImport/html/Dataset.html#method-format_dataset}{\code{GwasDataImport::Dataset$format_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="get_gwasdata_fields">}\href{../../GwasDataImport/html/Dataset.html#method-get_gwasdata_fields}{\code{GwasDataImport::Dataset$get_gwasdata_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="get_metadata_fields">}\href{../../GwasDataImport/html/Dataset.html#method-get_metadata_fields}{\code{GwasDataImport::Dataset$get_metadata_fields()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="is_new_id">}\href{../../GwasDataImport/html/Dataset.html#method-is_new_id}{\code{GwasDataImport::Dataset$is_new_id()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="se_from_bp">}\href{../../GwasDataImport/html/Dataset.html#method-se_from_bp}{\code{GwasDataImport::Dataset$se_from_bp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="set_wd">}\href{../../GwasDataImport/html/Dataset.html#method-set_wd}{\code{GwasDataImport::Dataset$set_wd()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="view_metadata_options">}\href{../../GwasDataImport/html/Dataset.html#method-view_metadata_options}{\code{GwasDataImport::Dataset$view_metadata_options()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="GwasDataImport" data-topic="Dataset" data-id="write_metadata">}\href{../../GwasDataImport/html/Dataset.html#method-write_metadata}{\code{GwasDataImport::Dataset$write_metadata()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$new(
  ebi_id,
  wd = tempdir(),
  ftp_path = NULL,
  igd_id = paste0("ebi-a-", ebi_id),
  traitname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id}}{e.g. GCST005522}

\item{\code{wd}}{Directory in which to download and develop dataset. Default=tempdir(). Deleted automatically upon object removal}

\item{\code{ftp_path}}{Pre-specified path to data. Default=NULL}

\item{\code{igd_id}}{Defaults to "ebi-a-<ebi_id>"}

\item{\code{traitname}}{Option to provide traitname of dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new EbiDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-download_dataset}{}}}
\subsection{Method \code{download_dataset()}}{
Download
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$download_dataset(
  ftp_path = self$ftp_path,
  ftp_url = options()$ebi_ftp_url,
  outdir = self$wd
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ftp_path}}{Pre-specified path to data. Default=self$ftp_path}

\item{\code{ftp_url}}{Default=options()$ebi_ftp_url}

\item{\code{outdir}}{Default=self$wd}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_ebi_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-format_ebi_dataset}{}}}
\subsection{Method \code{format_ebi_dataset()}}{
organise data before formatting. This is slow but doesn't really matter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$format_ebi_dataset(
  filename = self$filename,
  output = file.path(self$wd, "step1.txt.gz")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Filename of GWAS dataset}

\item{\code{output}}{Where to save formatted dataset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-organise_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-organise_metadata}{}}}
\subsection{Method \code{organise_metadata()}}{
Download and parse metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$organise_metadata(
  ebi_id = self$ebi_id,
  or_flag = self$or_flag,
  igd_id = self$igd_id,
  units = NULL,
  sex = "NA",
  category = "NA",
  subcategory = "NA",
  build = "HG19/GRCh37",
  group_name = "public",
  traitname = self$traitname
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id}}{Default=self$ebi_id}

\item{\code{or_flag}}{Default=self$or_flag}

\item{\code{igd_id}}{Default=NULL}

\item{\code{units}}{Default=NULL}

\item{\code{sex}}{Default="NA"}

\item{\code{category}}{Default="NA"}

\item{\code{subcategory}}{Default="NA"}

\item{\code{build}}{Default="HG19/GRCh37"}

\item{\code{group_name}}{Default="public"}

\item{\code{traitname}}{Default=self$traitname}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-pipeline}{}}}
\subsection{Method \code{pipeline()}}{
Once initialised this function will string together everything i.e. downloading, processing and uploading
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$pipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EbiDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
