% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.r
\name{get_positions}
\alias{get_positions}
\title{Lookup positions for given rsids in particular build}
\usage{
get_positions(rsid, build)
}
\arguments{
\item{rsid}{rsid}

\item{build}{build (36, 37 or 38)}
}
\value{
data frame
}
\description{
Lookup positions for given rsids in particular build
}
