% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.r
\name{listftp}
\alias{listftp}
\title{<brief desc>}
\usage{
listftp(url = options()$ebi_ftp_url, recursive = TRUE)
}
\arguments{
\item{url}{<what param does>}

\item{recursive=TRUE}{<what param does>}
}
\value{

}
\description{
<full description>
}
