% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain.r
\name{ObtainEbiDataset}
\alias{ObtainEbiDataset}
\title{Obtain EBI Dataset class}
\description{
Obtain EBI Dataset class

Obtain EBI Dataset class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObtainEbiDataset$new()}}
\item \href{#method-finalize}{\code{ObtainEbiDataset$finalize()}}
\item \href{#method-delete_wd}{\code{ObtainEbiDataset$delete_wd()}}
\item \href{#method-set_wd}{\code{ObtainEbiDataset$set_wd()}}
\item \href{#method-download_dataset}{\code{ObtainEbiDataset$download_dataset()}}
\item \href{#method-format_dataset}{\code{ObtainEbiDataset$format_dataset()}}
\item \href{#method-organise_metadata}{\code{ObtainEbiDataset$organise_metadata()}}
\item \href{#method-pipeline}{\code{ObtainEbiDataset$pipeline()}}
\item \href{#method-clone}{\code{ObtainEbiDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$new(
  ebi_id,
  wd = tempdir(),
  ftp_path = NULL,
  igd_id = paste0("ebi-a-", ebi_id),
  traitname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id}}{e.g. GCST004426}

\item{\code{igd_id}}{Defaults to "ebi-a-<ebi_id>"}

\item{\code{wd=tempdir()}}{<what param does>}

\item{\code{ftp_path=NULL}}{<what param does>}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new ObtainEbiDataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_wd"></a>}}
\if{latex}{\out{\hypertarget{method-delete_wd}{}}}
\subsection{Method \code{delete_wd()}}{
delete working directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$delete_wd()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_wd"></a>}}
\if{latex}{\out{\hypertarget{method-set_wd}{}}}
\subsection{Method \code{set_wd()}}{
set working directory (creates)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$set_wd(wd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wd}}{working directory}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-download_dataset}{}}}
\subsection{Method \code{download_dataset()}}{
Download
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$download_dataset(
  ftp_path = self$ftp_path,
  ftp_url = options()$ebi_ftp_url,
  outdir = self$wd
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ftp_path=self$ftp_path}}{<what param does>}

\item{\code{ftp_url=options()$ebi_ftp_url}}{<what param does>}

\item{\code{outdir=self$wd}}{<what param does>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-format_dataset}{}}}
\subsection{Method \code{format_dataset()}}{
format dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$format_dataset(dl = self$dl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dl=self$dl}}{<what param does>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-organise_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-organise_metadata}{}}}
\subsection{Method \code{organise_metadata()}}{
Download and parse metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$organise_metadata(
  ebi_id = self$ebi_id,
  or_flag = self$or_flag,
  igd_id = self$igd_id,
  units = NULL,
  sex = "NA",
  category = "NA",
  subcategory = "NA",
  build = "HG19/GRCh37",
  group_name = "public",
  traitname = self$traitname
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ebi_id=self$ebi_id}}{<what param does>}

\item{\code{or_flag=self$or_flag}}{<what param does>}

\item{\code{igd_id=NULL}}{<what param does>}

\item{\code{units=NULL}}{<what param does>}

\item{\code{sex="NA"}}{<what param does>}

\item{\code{category="NA"}}{<what param does>}

\item{\code{subcategory="NA"}}{<what param does>}

\item{\code{build="HG19/GRCh37"}}{<what param does>}

\item{\code{group_name="public"}}{<what param does>}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-pipeline}{}}}
\subsection{Method \code{pipeline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$pipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObtainEbiDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
